split provider wire code out of api.rs
api.rs is now just the provider-agnostic core (`Message`, `Role`). OpenAI's wire
types and `complete` move to a new openai.rs; the Anthropic-only
`field_or_placeholder` moves into anthropic.rs (and OpenAI no longer applies it
— it tolerates empty content). Also drops the unused OpenAI edits-API types
(EditRequest/EditResponse/Choice), which were dead.
